+++ views/Wallet/BalancePane.tsx
@@ -367,7 +367,6 @@ export default class BalancePane extends React.PureComponent<
                                             borderRadius: 10,
                                             borderColor:
                                                 themeColor('highlight'),
-                                            top: 10,
                                             margin: 20,
                                             padding: 15,
                                             borderWidth: 1.5
